move subscription ipc task processing to controller class
